Partial refactor (homepage, products)
