+++ gatsby-browser.js
@@ -1,4 +1,4 @@
-const { registerResolvers } = require('gatsby-source-prismic-graphql');
+const { registerLinkResolver } = require('gatsby-source-prismic-graphql');
 const { linkResolver } = require('./src/utils/linkResolver');
  
-registerResolvers(linkResolver);
+registerLinkResolver(linkResolver);
